+++ test/test.js
@@ -1659,6 +1659,31 @@ describe('Chat', () => {
 		});
 	});
 
+	describe('MML messages', () => {
+		describe('Error', () => {
+			it('Send invalid MML message', async () => {
+				const cmdChannel = authClient.channel('ai', 'excuse-test');
+				await cmdChannel.watch();
+				const cmd = '/mml-examples non-existing';
+				const data = await cmdChannel.sendMessage({ text: cmd });
+				expect(data.message.type).to.equal('error');
+			});
+		});
+
+		describe('Success', () => {
+			it('Send MML message', async () => {
+				const cmdChannel = authClient.channel('ai', 'excuse-test');
+				await cmdChannel.watch();
+				const cmd = '/mml-examples hi';
+				const data = await cmdChannel.sendMessage({ text: cmd });
+				expect(data.message.text).to.equal(cmd);
+				expect(data.message.mml).to.equal(
+					'\n\t\t<mml name="message">\n\t\t\t<text>Hi!</text>\n\t\t</mml>\n\t',
+				);
+			});
+		});
+	});
+
 	describe('Slash Commands', () => {
 		describe('Success', () => {
 			it('Giphy Integration', async () => {
@@ -1997,15 +2022,14 @@ describe('Chat', () => {
 		});
 
 		it('Typing Helpers', async () => {
-			let occurences = 0;
-			conversation.on('typing.start', () => {
-				occurences += 1;
-				if (occurences > 1) {
-					throw Error('too many typing.start events');
-				}
-			});
+			let occurrences = 0;
+			conversation.on('typing.start', () => occurrences++);
+
 			await conversation.keystroke();
 			await conversation.keystroke();
+
+			if (occurrences === 0) throw Error('typing.start never called');
+			if (occurrences > 1) throw Error('too many typing.start events');
 		});
 
 		it('Message Read', async () => {